+++ src/pages/Dir/Index.vue
@@ -22,6 +22,7 @@
             <q-input
               class="col-6"
               outlined
+              clearable
               label="Target URL"
               v-model="target"
               :hint="targetHint"
@@ -74,6 +75,7 @@
 <script lang="ts">
 import { defineComponent, provide, ref } from 'vue'
 import { useStore } from 'vuex'
+import { useRoute } from 'vue-router'
 import module from 'components/Module.vue'
 import actionBtn from 'components/Buttons/ActionBtn.vue'
 import scanRes from 'components/ScanRes.vue'
@@ -90,7 +92,6 @@ export default defineComponent({
     scanRes,
   },
   setup() {
-    const store = useStore()
     const targetHint = 'e.g. https://example.com; https://hackerone.com'
     const extHint = 'e.g. php,html,js; php,asp,aspx,jsp,bak,zip,tgz'
     const ext = ref('php')
@@ -103,7 +104,7 @@ export default defineComponent({
       args,
       form,
       formSubmit,
-    } = useScan(store)
+    } = useScan(useStore(), useRoute())
 
     const table = useTable(api, 'Directories')
     const { project_id_filter, getScans } = table