+++ src/pages/Domain/Index.vue
@@ -22,6 +22,7 @@
             <q-input
               class="col-6"
               outlined
+              clearable
               label="Domain name"
               v-model="target"
               :hint="targetHint"
@@ -60,6 +61,7 @@
 <script lang="ts">
 import { defineComponent, provide } from 'vue'
 import { useStore } from 'vuex'
+import { useRoute } from 'vue-router'
 import module from 'components/Module.vue'
 import actionBtn from 'components/Buttons/ActionBtn.vue'
 import scanRes from 'components/ScanRes.vue'
@@ -76,7 +78,6 @@ export default defineComponent({
     scanRes,
   },
   setup() {
-    const store = useStore()
     const targetHint = 'e.g. example.com; hackerone.com'
 
     const {
@@ -87,7 +88,7 @@ export default defineComponent({
       args,
       form,
       formSubmit,
-    } = useScan(store)
+    } = useScan(useStore(), useRoute())
 
     const table = useTable(api, 'Domains')
     const { project_id_filter, getScans } = table