+++ src/pages/Endpoint/Index.vue
@@ -22,6 +22,7 @@
             <q-input
               class="col-6"
               outlined
+              clearable
               label="Target URL"
               v-model="target"
               :hint="targetHint"
@@ -66,6 +67,7 @@
 <script lang="ts">
 import { defineComponent, provide, ref } from 'vue'
 import { useStore } from 'vuex'
+import { useRoute } from 'vue-router'
 import module from 'components/Module.vue'
 import actionBtn from 'components/Buttons/ActionBtn.vue'
 import scanRes from 'components/ScanRes.vue'
@@ -82,7 +84,6 @@ export default defineComponent({
     scanRes,
   },
   setup() {
-    const store = useStore()
     const targetHint = 'e.g. https://example.com; https://hackerone.com'
     const args_hak = ref('')
 
@@ -94,7 +95,7 @@ export default defineComponent({
       args,
       form,
       formSubmit,
-    } = useScan(store)
+    } = useScan(useStore(), useRoute())
 
     const table = useTable(api, 'Endpoints')
     const { project_id_filter, getScans } = table